requirements.txt
alembic==0.8.8
bcrypt==3.1.1
beautifulsoup4==4.5.1
blinker==1.4
cffi==1.8.3
click==6.6
factory-boy==2.7.0
fake-factory==0.7.2
Flask==0.11.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.0.0
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
psycopg2==2.6.2
py==1.4.31
pycparser==2.17
pytest==3.0.4
python-dateutil==2.6.0
python-editor==1.0.1
regex==2016.12.27
six==1.10.0
SQLAlchemy==1.1.3
waitress==1.0.1
WebOb==1.6.2
WebTest==2.0.23
Werkzeug==0.11.10
WTForms==2.1
